news-insight-backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
neo4j==5.14.0

# RSS & Parsing
feedparser==6.0.10
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
playwright>=1.40.0  # 동적 사이트 파싱 (fallback)
trafilatura>=1.6.0  # 지능형 본문 추출 (메타데이터 포함)

# AI
openai>=1.10.0

# Embedding & Similarity
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
torch>=2.0.0  # sentence-transformers 및 KoBART 요약에 필요

# KeyBERT (의미 기반 키워드 추출)
keybert>=0.8.0

# Kiwi (한국어 형태소 분석 및 복합명사 추출)
kiwipiepy>=0.18.0

# KSS (한국어 문장 분할)
kss>=6.0.0

# TextRank (문장 추출)
summa>=1.2.0
networkx>=3.0

# KoBART (요약 생성) - transformers는 KoBART 요약에 필요
transformers>=4.35.0
sentencepiece>=0.1.99
accelerate>=0.24.0
safetensors>=0.4.0

# Gemini API
google-generativeai>=0.3.0

# Celery & Redis
celery==5.3.4
redis==5.0.1

# 기업명 데이터 수집 (선택적 - 스크립트 실행 시만 필요)
pykrx>=1.0.0  # 한국 기업명 수집 (최신 버전: 1.0.51)
finance-datareader>=0.9.50  # 한국 주식 섹터 정보 수집 (FDR)
pandas>=2.0.0  # 미국 기업명 수집
finnhub-python>=2.4.18  # Finnhub API (미국 기업명 수집)

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
psutil>=5.9.0  # 프로세스 관리 (중복 실행 감지용)

# Logging
python-json-logger==2.0.7

# Retry & Resilience
tenacity>=8.2.3  # Exponential backoff, retry 로직

# JSON Repair
json-repair>=0.7.0  # JSON 자동 수리

# DART API
OpenDartReader>=0.1.6  # DART API 클라이언트

# Markdown
markdownify>=0.11.6  # HTML → Markdown 변환

# Embedding
FlagEmbedding>=1.2.0  # BGE-M3 임베딩 모델

# LangChain (LLM 호출)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1  # MarkdownHeaderTextSplitter

# PDF Parsing (Phase 2)
pdfplumber>=0.10.0  # Smart PDF Parser (2단 레이아웃 처리)
rapidfuzz>=3.0.0  # Fuzzy string matching (Ticker 매칭용)

